requirements.txt
matplotlib==3.9.0
numpy==1.26.4
Pillow==10.3.0
torch==2.3.1
torchvision==0.18.1
